requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Code Execution & Analysis
human-eval>=1.0.3
pytest>=7.4.0
ast-comments>=1.1.0
pycodestyle>=2.11.0
radon>=6.0.1  # Cyclomatic complexity

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
wandb>=0.15.0  # Experiment tracking (optional)

# Development
jupyter>=1.0.0
ipython>=8.12.0
black>=23.3.0  # Code formatting
flake8>=6.0.0  # Linting

# Testing
pytest-cov>=4.1.0
pytest-xdist>=3.3.0  # Parallel testing